+++ modules/ustk_core/src/usImageRF3D.cpp
@@ -32,7 +32,7 @@ usImageRF3D::usImageRF3D(unsigned int AN, unsigned int LN, unsigned int FN)
 * @param[in] isImageConvex Boolean to specyfy if the image was acquired by a convex probe(true) or by a linear probe (false).
 * @param[in] isMotorConvex Boolean to specyfy if the image was acquired by a rotating  motor(true) or by a linear motor (false).
 */
-usImageRF3D::usImageRF3D(unsigned int AN, unsigned int LN, unsigned int FN, float probeRadius, float motorRadius, float scanLinePitch, float framePitch,
+usImageRF3D::usImageRF3D(unsigned int AN, unsigned int LN, unsigned int FN, double probeRadius, double motorRadius, double scanLinePitch, double framePitch,
   bool isImageConvex, bool isMotorConvex)
   : usImage3D<short>(AN, LN, FN), usImageSettings3D(probeRadius, motorRadius, scanLinePitch, framePitch, isImageConvex, isMotorConvex)
 {
@@ -106,10 +106,10 @@ unsigned int  usImageRF3D::getFN() const { return getDimZ(); }
 * Setter for axial Resolution.
 * @param Axial resolution (in meters) to set.
 */
-void usImageRF3D::setAxialResolution(float axialResolution) { m_axialResolution = axialResolution; }
+void usImageRF3D::setAxialResolution(double axialResolution) { m_axialResolution = axialResolution; }
 
 /**
 * Getter for axial Resolution.
 * @return Axial resolution (in meters).
 */
-float usImageRF3D::getAxialResolution() const { return m_axialResolution; }
+double usImageRF3D::getAxialResolution() const { return m_axialResolution; }
